com.cubrid.cubridmanager.core.testfragment/src/com/cubrid/cubridmanager/core/cubrid/table/task/GetAutoIncrementTaskTest.java
/*
 * Copyright (C) 2009 Search Solution Corporation. All rights reserved by Search Solution. 
 *
 * Redistribution and use in source and binary forms, with or without modification, 
 * are permitted provided that the following conditions are met: 
 *
 * - Redistributions of source code must retain the above copyright notice, 
 *   this list of conditions and the following disclaimer. 
 *
 * - Redistributions in binary form must reproduce the above copyright notice, 
 *   this list of conditions and the following disclaimer in the documentation 
 *   and/or other materials provided with the distribution. 
 *
 * - Neither the name of the <ORGANIZATION> nor the names of its contributors 
 *   may be used to endorse or promote products derived from this software without 
 *   specific prior written permission. 
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
 * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
 * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. 
 * IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
 * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
 * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, 
 * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
 * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY 
 * OF SUCH DAMAGE. 
 *
 */
package com.cubrid.cubridmanager.core.cubrid.table.task;

import java.util.ArrayList;

import com.cubrid.common.core.common.model.SerialInfo;
import com.cubrid.cubridmanager.core.SetupJDBCTestCase;

/**
 * test getting auto increment information
 * 
 * @author moulinwang
 * @version 1.0 - 2009-7-1 created by moulinwang
 */
public class GetAutoIncrementTaskTest extends
		SetupJDBCTestCase {
	String testTableName = "testGetAutoIncrementTaskTest";
	String sql = null;

	private boolean createTestTable() {
		String sql = "create table \"" + testTableName + "\" ("
				+ "auto smallint AUTO_INCREMENT(3,5))";
		return executeDDL(sql);
	}

	private boolean dropTestTable() {
		String sql = "drop table \"" + testTableName + "\"";
		return executeDDL(sql);
	}

	public void testGetAutoIncrementTaskTest() {
		createTestTable();
		//test normal case
		GetAutoIncrementTask task = new GetAutoIncrementTask(databaseInfo);
		task.setTableName(testTableName);
		task.execute();
		ArrayList<SerialInfo> list = task.getSerialInfoList();
		assertEquals(1, list.size());
		SerialInfo serial = list.get(0);
		assertEquals("3", serial.getMinValue());
		assertEquals("5", serial.getIncrementValue());

		//test exception case: connection is closed
		task.execute();

		//test exception case: have error
		task.setErrorMsg("error");
		task.execute();
	}

	@Override
	protected void tearDown() throws Exception {
		dropTestTable();
		super.tearDown();
	}
}